split up fixtures and nwb fixture in particular, add --clean pytest option
