--- nwb_models/src/nwb_models/models/pydantic/__init__.py
@@ -1,1 +1,0 @@
-from .pydantic.core.v2_7_0.namespace import *
